refactor: fcmController에 있는 테스트 API TestController로 이동
